+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Allow to customize ID Pod selectors
 - Add Label and Envs Pod customization
 - Improved JWT Rotation
+- Allow to customize Security Context in pods
 
 ## [1.0.3](https://github.com/arangodb/kube-arangodb/tree/1.0.3) (2020-05-25)
 - Prevent deletion of not known PVC's